requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-cors-headers==3.0.2
django-filter==2.2.0
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
drf-nested-routers==0.91
gunicorn==19.9.0
idna==2.8
Markdown==3.1.1
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
